pyproject.toml
[project]
name = "markov-agent"
version = "0.1.0"
description = "A topology-based probabilistic control system using Google ADK."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.1.0",           # The Agent Engine
    "pydantic>=2.7.0",             # The Structure
    "tenacity>=8.2.0",             # Retry logic
    "jinja2>=3.1.0",               # Prompt Templating
    "numpy>=1.26.0",               # Math/Stats for simulation
    "rich>=13.0.0",                # Terminal visualization
    "litellm>=1.80.13",
    "sqlalchemy>=2.0.45",
    "httpx>=0.28.1",
    "mcp>=1.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",                 # The Linter & Formatter
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "D100",   # Missing docstring in public module
    "D101",   # Missing docstring in public class
    "D102",   # Missing docstring in public method
    "D103",   # Missing docstring in public function
    "D104",   # Missing docstring in public package
    "D107",   # Missing docstring in __init__
    "ANN101", # Missing type annotation for self
    "ANN102", # Missing type annotation for cls
    "ANN204", # Missing return type annotation for special method
    "S101",   # Use of assert detected (needed for tests)
    "PLR2004",# Magic value used in comparison
    "ARG005", # Unused lambda argument
    "PLC0415",# Import should be at the top-level
    "BLE001", # Do not catch blind exception
    "COM812", # Trailing comma missing (handled by formatter)
    "E501",   # Line too long
    "INP001", # Implicit namespace package
]

extend-select = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "D", "S", "PLR", "ARG", "FBT", "SLF"]
"examples/*" = ["ANN", "D", "T201", "PLR", "ARG", "FBT", "BLE", "PTH", "S", "TRY", "EM", "PERF", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
